fall-2023/week-3/twitter-scraper/requirements.txt
beautifulsoup4
NwalaTextUtils
pytest-playwright
requests
urllib3<2.0